fixed #32 - empty repository creation failure.
Creating an empty repository with branch protection rules causes the api to respond with an error, as there are no files in the repository and therefore no branch. Work around implemented which either adds a default README.md file if there are branch-protection rules defined. If not, we have to remove the default_branch parameter from the payload as it introduces the same issue again.

Signed-off-by: Tim Beermann <beermann@osism.tech>